# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FiveM Casino API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /slots/banker-settings:
    get:
      operationId: getSlotsSettings
      tags: [slots]
      summary: Get slot machine settings (banker only)
      responses:
        "200":
          description: Slot settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlotsSettings"
    post:
      operationId: updateSlotsSettings
      tags: [slots]
      summary: Update slot machine settings (banker only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlotsSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlotsSettings"
  /slots/spin:
    post:
      operationId: spinSlots
      tags: [slots]
      summary: Spin the slot machine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bet:
                  type: integer
              required: [bet]
      responses:
        "200":
          description: Spin result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlotsSpinResult"
...
components:
  schemas:
...
    SlotsSettings:
      type: object
      properties:
        enabled:
          type: boolean
        minBet:
          type: integer
        maxBet:
          type: integer
        fortunaBetSteps:
          type: array
          items:
            type: integer
        westernSlotsBetSteps:
          type: array
          items:
            type: integer
      required: [enabled, minBet, maxBet]
    SlotsSpinResult:
      type: object
      properties:
        reels:
          type: array
          items:
            type: string
        multiplier:
          type: number
        payout:
          type: integer
        outcome:
          type: string
          enum: [jackpot, big_win, win, small_win, lose]
        playerChips:
          type: integer
      required: [reels, multiplier, payout, outcome, playerChips]
...